refactor: reParseSearch
